renaming SingleTenantEntityReference.java
